+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 * Fixed issue where correct error message would not propagate for invalid images.
+* Added the *--description* option to the **init** command.
+* Added the *--contribution* option to the **init** command which converts a contribution zip to proper pack format.
 * Improved **validate** command performance time and outputs.
 * Added the flag *--no-docker-checks* to **validate** command to skip docker checks.
 * Added the flag *--print-ignored-files* to **validate** command to print ignored files report when the command is done.
@@ -10,6 +12,7 @@
    * Validating that the timestamp in the "created" field in the pack_metadata is in ISO format.
 * Running `demisto-sdk validate` will run the **validate** command using git and only on committed files (same as using *-g --post-commit*).
 * Fixed an issue where release notes were not checked correctly in **validate** command.
+* Fixed an issue in the **create-id-set** command where optional playbook tasks were not taken into consideration.
 
 #### 1.1.5
 * Fixed an issue in **find-dependencies** command.